Rename psvDebugScreen* header file to debugScreen
